pyproject.toml
[project]
name = "shift-manager-bot"
version = "0.1.0"
description = ""
authors = [{ name = "Komron", email = "alexhirsch293@gmail.com" }]
readme = "README.md"
requires-python = ">=3.14,<3.15"
dependencies = [
  "aiogram (>=3.25.0,<4.0.0)",
  "fastapi (>=0.132.0,<0.133.0)",
  "uvicorn (>=0.41.0,<0.42.0)",
  "sqlalchemy (>=2.0.46,<3.0.0)",
  "asyncpg (>=0.31.0,<0.32.0)",
  "alembic (>=1.18.4,<2.0.0)",
  "redis (>=7.2.0,<8.0.0)",
  "apscheduler (>=3.11.2,<4.0.0)",
  "pydantic-settings (>=2.13.1,<3.0.0)",
  "python-dotenv (>=1.2.1,<2.0.0)",
  "greenlet (>=3.3.2,<4.0.0)",
  "telegram-init-data[fastapi] (>=1.0.2,<2.0.0)",
]

[tool.poetry]
packages = [{ include = "shift_manager_bot", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
  "pytest (>=9.0.2,<10.0.0)",
  "pytest-asyncio (>=1.3.0,<2.0.0)",
  "pytest-mock (>=3.15.1,<4.0.0)",
  "factory-boy (>=3.3.3,<4.0.0)",
  "httpx (>=0.28.1,<0.29.0)",
  "aiogram-tests (>=1.0.3,<2.0.0)",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.9
target-version = "py39"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F401", "F841"]
"alembic/*" = ["F401"]
"src/shift_manager_bot/migrations/*" = ["F401"]
